refactor: move common components
