+++ src/components/Landing/HeroSection/index.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import Video from '../../../videos/video.mp4'
-import { Button } from '../../ButtonElements'
+import { Button } from '../../Common/Button'
 import {
   HeroContainer,
   HeroBg,
