+++ src/components/Landing/infoSection/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Button } from '../../ButtonElements'
+import { Button } from '../../Common/Button'
 import {
   InfoContainer,
   InfoWrapper,
